更新了 6-6 输入捕获模式测频率 ， 修复了一个错误， 删除了一个冗余引用。

+++ 6-6 输入捕获模式测频率/Hardware/IC.h
@@ -3,6 +3,6 @@
 
 void IC_Init(void);
 uint32_t IC_GetFreq(void);
-uint8_t IC_GetCapture1(void);
+uint16_t IC_GetCapture1(void);
 
 #endif
